# codegen/swagger.yaml
swagger: "2.0"
info:
  description: "D3 mentoring test application"
  version: "1.0.0"
  title: "Mentoring D3 Swagger"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "mikhail_kipets@epam.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/v2"
...
schemes:
- "https"
- "http"
paths:
  /file:
    post:
      tags:
      - "file"
      summary: "Add a new file to the storage"
      description: ""
      operationId: "addFile"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "File object that needs to be added to the store"
        required: true
        schema:
          $ref: "#/definitions/File"
      responses:
        405:
          description: "Invalid input"
      security:
      - storage_auth:
        - "write:files"
        - "read:files"
    put:
      tags:
      - "file"
      summary: "Update an existing file"
      description: ""
      operationId: "updateFile"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "File object that needs to be added to the store"
        required: true
        schema:
          $ref: "#/definitions/File"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "File not found"
        405:
          description: "Validation exception"
      security:
      - storage_auth:
        - "write:files"
        - "read:files"
  /file/findByStatus:
    get:
      tags:
      - "file"
      summary: "Finds Files by status"
      description: "Multiple status values can be provided with comma separated strings"
      operationId: "findFilesByStatus"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "status"
        in: "query"
        description: "Status values that need to be considered for filter"
        required: true
        type: "array"
        items:
          type: "string"
          enum:
          - "available"
          - "pending"
          - "outdated"
          default: "available"
        collectionFormat: "multi"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/File"
        400:
          description: "Invalid status value"
      security:
      - storage_auth:
        - "write:files"
        - "read:files"
  /file/findByTags:
    get:
      tags:
      - "file"
      summary: "Finds Files by tags"
      description: "Muliple tags can be provided with comma separated strings. Use         tag1, tag2, tag3 for testing."
      operationId: "findFilesByTags"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "tags"
        in: "query"
        description: "Tags to filter by"
        required: true
        type: "array"
        items:
          type: "string"
        collectionFormat: "multi"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/File"
        400:
          description: "Invalid tag value"
      security:
      - storage_auth:
        - "write:files"
        - "read:files"
      deprecated: true
  /file/{fileId}:
    get:
      tags:
      - "file"
      summary: "Find file by ID"
      description: "Returns a single file"
      operationId: "getFileById"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "fileId"
        in: "path"
        description: "ID of file to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/File"
        400:
          description: "Invalid ID supplied"
        404:
          description: "File not found"
      security:
      - api_key: []
    post:
      tags:
      - "file"
      summary: "Updates a file in the store with form data"
      description: ""
      operationId: "updateFileWithForm"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "fileId"
        in: "path"
        description: "ID of file that needs to be updated"
        required: true
        type: "integer"
        format: "int64"
      - name: "name"
        in: "formData"
        description: "Updated name of the file"
        required: false
        type: "string"
      - name: "status"
        in: "formData"
        description: "Updated status of the file"
        required: false
        type: "string"
      responses:
        405:
          description: "Invalid input"
      security:
      - storage_auth:
        - "write:files"
        - "read:files"
    delete:
      tags:
      - "file"
      summary: "Deletes a file"
      description: ""
      operationId: "deleteFile"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "api_key"
        in: "header"
        required: false
        type: "string"
      - name: "fileId"
        in: "path"
        description: "File id to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "File not found"
      security:
      - storage_auth:
        - "write:files"
        - "read:files"
  /file/{fileId}/uploadImage:
    post:
      tags:
      - "file"
      summary: "uploads an image"
      description: ""
      operationId: "uploadFile"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "fileId"
        in: "path"
        description: "ID of file to update"
        required: true
        type: "integer"
        format: "int64"
      - name: "additionalMetadata"
        in: "formData"
        description: "Additional data to pass to server"
        required: false
        type: "string"
      - name: "file"
        in: "formData"
        description: "file to upload"
        required: false
        type: "file"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      security:
      - storage_auth:
        - "write:files"
        - "read:files"
...
securityDefinitions:
  storage_auth:
    type: "oauth2"
    authorizationUrl: "http://localhost:8080/oauth/dialog"
    flow: "implicit"
    scopes:
      write:files: "modify pfiles in your account"
      read:files: "read your files"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  Category:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
    xml:
      name: "Category"
...
  Tag:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
    xml:
      name: "Tag"
  File:
    type: "object"
    required:
    - "name"
    - "photoUrls"
    properties:
      id:
        type: "integer"
        format: "int64"
      category:
        $ref: "#/definitions/Category"
      name:
        type: "string"
        example: "doggie"
      photoUrls:
        type: "array"
        xml:
          name: "photoUrl"
          wrapped: true
        items:
          type: "string"
      tags:
        type: "array"
        xml:
          name: "tag"
          wrapped: true
        items:
          $ref: "#/definitions/Tag"
      status:
        type: "string"
        description: "file status in the storage"
        enum:
        - "available"
        - "pending"
        - "outdated"
    xml:
      name: "File"
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      type:
        type: "string"
      message:
        type: "string"
...